# doc/wanted-board.swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: wanted board
servers:
  - url: http://localhost:3100/v1/api
paths:
  /posts:
    post:
      description: 게시글 등록
      tags:
        - post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistPostRequest'
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistPostResponse'
    get:
      description: 게시글 목록 조회
      tags:
        - post
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: string
        - name: title
          in: query
          required: false
          schema:
            type: string
        - name: writer
          in: query
          required: false
          schema:
            type: string   
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/FetchPostResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /posts/{id}:
    delete: 
      description: 게시글 삭제
      tags:
        - post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: password
          in: header
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: 
                    type: integer
                    example: 1
    put:
      description: 게시글 수정
      tags:
        - post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: password
          in: header
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        200:
          description: OK
  /replies:
    post:
      description: 댓글 등록
      tags:
        - reply
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistReplyRequest'
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistReplyResponse'
  /replies/{post_id}:
    get:
      description: 댓글 목록 조회
      tags:
        - reply
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  replies:
                    type: array
                    items:
                      $ref: '#/components/schemas/FetchReplyResponse'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
components: 
  schemas:
    RegistPostRequest:
      type: object
      properties:
        title:
          type: string
          example: rain
        content:
          type: string
          example: I love rainy day
        writer:
          type: string
          example: rain man
        password:
          type: string
          example: qwe123
    RegistPostResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: rain
        writer:
          type: string
          example: rain man
        created_at:
          type: string
          format: date-time
          example: 2022-05-30T11:44:02.297Z
    FetchPostResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: rain
        content:
          type: string
          example: I love rainy day
        writer:
          type: string
          example: rain man
        created_at:
          type: string
          format: date-time
          example: 2022-05-30T09:35:50.958Z
        updated_at:
          type: string
          format: date-time
          example: 2022-05-30T09:35:50.958Z
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          example: rain
        content:
          type: string
          example: I love rainy day
    RegistReplyRequest:
      type: object
      properties:
        post_id:
          type: integer
          example: 1
        content:
          type: string
          example: Shall we dance in the rainy day?
        writer:
          type: string
          example: dance machine
        is_child:
          type: boolean
          example: false
        parent_id:
          type: string
          nullable: true
          example: null
    RegistReplyResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        post_id:
          type: integer
          example: 1
        writer:
          type: string
          example: dance machine
        is_child:
          type: boolean
          example: false
        parent_id:
          type: integer
          example: null
        create_at:
          type: string
          format: date-time
          example: 2022-05-30T15:54:04.866Z
    FetchChildReplyResponse:
      type: object
      properties:
        id:
          type: integer
          example: 2
        content:
          type: string
          example: I love dance!
        created_at:
          type: string
          format: date-time
          example: 2022-05-30T09:35:51.004Z
    FetchReplyResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        post_id:
          type: integer
          example: 1
        content:
          type: string
          example: Shall we dance in the rainy day?
        created_at:
          type: string
          format: date-time
          example: 2022-05-30T09:35:51.004Z
        child_reply:
          $ref: '#/components/schemas/FetchChildReplyResponse'
    Pagination:
      type: object
      properties:
        page: 
          type: integer
          example: 1
        limit: 
          type: integer
          example: 500
        total_count:
          type: integer
          example: 5